moved Filter component to src/components folder
